+++ server/odc-migrate/src/main/resources/init-config/runtime/iam_user_role_individual_space.yaml
@@ -1,0 +1,22 @@
+kind: resource
+version: v2
+templates:
+    - metadata:
+          allow_duplicate: false
+          table_name: iam_user_role
+          unique_keys: [ "role_id", "user_id" ]
+      specs:
+          - column_name: user_id
+            value: ${CREATOR_ID}
+            data_type: java.lang.Long
+          - column_name: role_id
+            value_from:
+                field_ref:
+                    ref_file: migrate/web/V_4_2_1_2__iam_role.yaml
+                    field_path: templates.0.specs.0.value
+          - column_name: organization_id
+            value: ${ORGANIZATION_ID}
+            data_type: java.lang.Long
+          - column_name: creator_id
+            value: ${CREATOR_ID}
+            data_type: java.lang.Long
