+++ server/odc-service/src/main/java/com/oceanbase/odc/service/collaboration/OrganizationResourceMigrator.java
@@ -13,105 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package com.oceanbase.odc.service.collaboration;
 
-import java.util.Arrays;
-import java.util.Objects;
-import java.util.Optional;
-
-import javax.sql.DataSource;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.support.TransactionTemplate;
-
-import com.oceanbase.odc.core.authority.util.SkipAuthorize;
-import com.oceanbase.odc.core.shared.constant.OrganizationType;
-import com.oceanbase.odc.metadb.iam.OrganizationEntity;
-import com.oceanbase.odc.service.iam.OrganizationService;
-import com.oceanbase.odc.service.iam.VerticalPermissionValidator;
-import com.oceanbase.odc.service.iam.model.Organization;
 import com.oceanbase.odc.service.iam.model.User;
 
-import lombok.extern.slf4j.Slf4j;
-
-/**
- * @Author: Lebie
- * @Date: 2023/5/30 10:53
- * @Description: []
- */
-@Service
-@Slf4j
-@SkipAuthorize
-public class OrganizationResourceMigrator {
-
-    @Autowired
-    private DataSource dataSource;
-
-    @Autowired
-    private IndividualOrganizationMigrator individualOrganizationMigrator;
-
-    @Autowired
-    private TeamOrganizationMigrator teamOrganizationMigrator;
-
-    @Autowired
-    private OrganizationService organizationService;
-
-    @Autowired
-    private RuleApplyingMigrator ruleApplyingMigrator;
-
-    @Autowired
-    private TransactionTemplate transactionTemplate;
-
-    @Autowired
-    private VerticalPermissionValidator verticalPermissionValidator;
-
-    public void migrate(User user) {
-        if (Objects.isNull(user)) {
-            return;
-        }
-        this.transactionTemplate.execute((transactionStatus -> {
-            try {
-                Optional<Organization> organization = organizationService.getByOrganizationTypeAndUserId(
-                        OrganizationType.TEAM, user.getId());
-                if (!organization.isPresent()) {
-                    log.info(
-                            "First user login to team organization, start to initialize resource, organizationId={}, userId={}",
-                            user.getOrganizationId(), user.getId());
-                    teamOrganizationMigrator.migrate(user);
-                    log.info("Initialized team organization resource successfully, organizationId={}, userId={}",
-                            user.getOrganizationId(), user.getId());
-                    ruleApplyingMigrator.migrate(user.getOrganizationId());
-                }
-                return null;
-            } catch (Exception ex) {
-                transactionStatus.setRollbackOnly();
-                throw new IllegalStateException(ex);
-            }
-        }));
-        this.transactionTemplate.execute((transactionStatus -> {
-            try {
-                Organization team = organizationService
-                        .getByOrganizationTypeAndUserId(OrganizationType.TEAM, user.getId())
-                        .orElseThrow(() -> new RuntimeException("User doesn't belong to any TEAM organization"));
-                Organization individual = new Organization();
-                individual.setId(-1L);
-                individual.setType(OrganizationType.INDIVIDUAL);
-                boolean hasIndividualPermission =
-                        verticalPermissionValidator.implies(individual, Arrays.asList("read", "update"), team.getId());
-                Optional<Organization> organization = organizationService.getByOrganizationTypeAndUserId(
-                        OrganizationType.INDIVIDUAL, user.getId());
-                if (!organization.isPresent() && hasIndividualPermission) {
-                    log.info("Individual organization not found, start to initialize for userId={}", user.getId());
-                    OrganizationEntity saved = individualOrganizationMigrator.migrate(user);
-                    log.info("Initialized individual organization successfully for userId={}", user.getId());
-                    ruleApplyingMigrator.migrate(saved.getId());
-                }
-                return null;
-            } catch (Exception ex) {
-                transactionStatus.setRollbackOnly();
-                throw new IllegalStateException(ex);
-            }
-        }));
-    }
+public interface OrganizationResourceMigrator {
+    void migrate(User user);
 }
